Add shared-memory and posix-mq interfaces to fix background blur
As suggested by Peter Bui <pbui@bx612.space>, add a shared-memory plug
to allow the Zoom snap to create the files like

    /dev/shm/aomshm.337bce.0

that it uses internally to communicated with the background blur helper
processes it creates.

Also add a dummy posix-mq slot to allow Zoom to create and manage the
POSIX message queues it uses for communication with the helper process.

This should fully fix issue #128.